### 2020-11-21 #2856 @whedon
[PRE REVIEW]: Isoreader: An R package to read stable isotope data files for reproducible research
**Submitting author:** @sebkopf (<a href="http://orcid.org/0000-0002-2044-0201">Sebastian Kopf</a>)
**Repository:** <a href="https://github.com/isoverse/isoreader" target ="_blank">https://github.com/isoverse/isoreader</a>
**Version:** v1.2.7
