### 2020-11-23 #2856 @mikldk
@sebkopf Thank you for your submission. I have a few questions and comments:

1) Please remove the `https://doi.org/` prefix in [paper.bib#L22](https://github.com/isoverse/isoreader/blob/master/paper.bib#L22)
2) Please include countries where appropriate in the paper (currently only the Netherlands is mentioned, the US is not)
3) Please confirm the paper's authors; when I look at the repo's [insights](https://github.com/isoverse/isoreader/graphs/contributors) I (almost) only see you as a developer (also refer to [JOSS's authorship docs](https://joss.readthedocs.io/en/latest/submitting.html#authorship))
4) Do you have any suggestions for reviewers (find the list above)?
