### 2020-11-23 524330b
joss submission fixes
- fix DOI
- add country specification to all affiliations
