### 2020-11-23 #2856 @sebkopf
@mikldk thank you very much for agreeing to be the editor for this submission

> Please remove the https://doi.org/ prefix in paper.bib#L22
> Please include countries where appropriate in the paper (currently only the Netherlands is mentioned, the US is not)

DOI issue and missing country in affiliations addressed in commit https://github.com/isoverse/isoreader/commit/524330bcdc8573ec796641c8751273545f02fdb7

> Please confirm the paper's authors; when I look at the repo's insights I (almost) only see you as a developer (also refer to JOSS's authorship docs)

I confirm that the paper's authors are correct. You are right that I did almost all of the actual coding but the biggest challenges for this R package were analyzing and testing different vendor file formats to catch data files that caused trouble and figure out exactly what information was stored where in binary data files. The two co-authors contributed enormously to these aspects of the project, which I consider to merit co-authorship if that's okay with JOSS policy.

> Do you have any suggestions for reviewers (find the list above)?

It seems there are not many reviewers (yet) in the field of geochemistry, I tried to find a few other potential reviewers who work on data access, ecology or reproducible research in R. Hopefully one of the following will be willing to review this submission.

 - jhollist
 - r-barnes
 - ethanwhite
 - kdorheim
 - dvanic

Thank you
