### 2020-11-24 #2856 @kdorheim
@mikldk thanks for thinking of me, I can review this submission to JOSS. Should the review be included in this issue or a separate issue? Also is there an expected due date for this review? Thanks!
