### 2020-11-24 08322d0
update readme
- add link to Scan vignette
- clarify that testing is via GitHub actions now (no longer Travis and AppVeyor)
