+++ pkgdown.yml
@@ -8,7 +8,7 @@ articles:
   operations: operations.html
   quick_start: quick_start.html
   scan: scan.html
-last_built: 2020-11-24T19:35Z
+last_built: 2020-11-24T21:40Z
 urls:
   reference: https://isoreader.isoverse.org//reference
   article: https://isoreader.isoverse.org//articles
