### 2020-11-25 #2856 @mikldk
> @mikldk thanks for thinking of me, I can review this submission to JOSS.
Thanks!

> Should the review be included in this issue or a separate issue?
It's up to you. I normally recommend specific comments as issues in the repository (linked to the review issue once created), and general comments/question in the review issue (once created).

> Also is there an expected due date for this review? Thanks!
No, but we normally try to get thinks done within weeks if possible - but if it's not, that's okay, too!
