### 2020-11-30 #2856 @sebkopf
> @sebkopf :
>
> Thanks.
>
> * [ ]  Please also add countries in the Acknowledgements section.

Country information added to Acknowledgements https://github.com/isoverse/isoreader/commit/9dae60c6277be1b2eb727111513d6c0ceface3ea. Thank you very much @kdorheim for reviewing our work.
