### 2020-12-03 #2878 @whedon
[REVIEW]: Isoreader: An R package to read stable isotope data files for reproducible research
**Submitting author:** @sebkopf (<a href="http://orcid.org/0000-0002-2044-0201">Sebastian Kopf</a>)
**Repository:** <a href="https://github.com/isoverse/isoreader" target ="_blank">https://github.com/isoverse/isoreader</a>
**Version:** v1.3.0

@kdorheim and @r-barnes, please carry out your review in this issue by updating the checklist below. If you cannot edit the checklist please:
