### 2020-12-04 isoverse/isoreader#141 @kdorheim
JOSS plain text license files
Related to: openjournals/joss-reviews/issues/2878

_Does the repository contain a plain-text LICENSE file with the contents of an OSI approved software license?_

Although GPL (>= 2) is defined as the license in the DESCRIPTION file I think that adding the text from https://opensource.org/licenses/GPL-2.0 to a LICENSE file would suffice this JOSS requirement.
