### 2020-12-04 isoverse/isoreader#142 @kdorheim
JOSS Contribution and authorship
_Contribution and authorship: Has the submitting author (@sebkopf) made major contributions to the software? Does the full list of paper authors seem appropriate and complete?_

Related to: openjournals/joss-reviews/issues/2878

There are [three coauthors](https://github.com/openjournals/joss-papers/blob/joss.02878/joss.02878/10.21105.joss.02878.pdf) listed on the paper but only one in the DESCRIPTION file. If coauthors Kroll and Kocken contributed to the package they should be listed there as well. If these authors only contributed to the paper text then the _Contribution and authorship_ task is complete.
