### 2020-12-04 isoverse/isoreader#143 @kdorheim
JOSS Installation instructions
_Installation instructions: Is there a clearly-stated list of dependencies? Ideally these should be handled with an automated package management solution._

Related to: openjournals/joss-reviews/issues/2878

The installation instructions were great installing from CRAN and github worked well and the instructions for troubleshooting were useful. However, the dependencies are not version controlled, at a minimum I'd suggest adding the package versions to NAMESPACE (as you did with `tidyselect (>= 1.0.0)`) or possibly use [renv](https://rstudio.github.io/renv/articles/renv.html) (speaking from personal experience building R packages that depend on tidyverse packages, which at times can have unstable versioning, renv has been a life saver).
