### 2020-12-04 isoverse/isoreader#144 @kdorheim
JOSS Automated tests
Related to: openjournals/joss-reviews/issues/2878

Isoreader does have automated testing that is deployed continuously via github actions 👏. Locally when I run devtools::test()

```
OK:       1241
Failed:   0
Warnings: 11
Skipped:  0
```
there are several warnings that were triggered, none of these seem to be alarming more of an artifact related to the way the tests were set up using random numbers. Could you confirm that none of these warnings are trival? Would it make sense to add documentation or suppress the warnings?
