### 2020-12-04 isoverse/isoreader#145 @kdorheim
JOSS Example usage
_Example usage: Do the authors include examples of how to use the software (ideally to solve real-world analysis problems)._

Related to: openjournals/joss-reviews/issues/2878

The vignettes are well laid out and demonstrate the power of this package. Working my way through these examples they all work but [here](https://isoreader.isoverse.org/articles/operations.html) they mention `iso_turn_info_message_off()` which I have been unable to use or locate the code for it, was this functional at some point but since then been depreciated?
