### 2020-12-04 #2878 @kdorheim
@sebkopf, `isoreader` is well written, well documented, and fulfills an important need in supporting reproducible open-source science. As indicated by the manuscripts that have cited this package it has already been useful to the community. The paper clearly describes the need for this package and is well written for a general audience (as a non-specialist I really appreciated that), however, it seems like the authors could have referenced citations for some of the information in the summary (i.e. the comments about availability and variability of isotopes). This being said if these generally known facts in the community, then no citations are needed. I've opened some issues on the `isoreader` repo related to the [JOSS](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html) guidelines, which I believe are relatively minor fixes.
