### 2020-12-06 #2856 @r-barnes
@mikldk: I've been a little slow with my reviews lately, but I could probably find time. The fact that this is already on CRAN means that it's passed a series of peer-reviews already with respect to various aspects of code quality, so a priori I don't expect too much difficulty.
