### 2020-12-14 #2878 @kdorheim
Hi @mikldk, I've gone through the checklist, opened issues when relevant/marked others as complete.

> JOSS plain text license files isoverse/isoreader#141
 Open
JOSS Contribution and authorship isoverse/isoreader#142
 Open
JOSS Installation instructions isoverse/isoreader#143
 Open
JOSS Automated tests isoverse/isoreader#144
 Open
JOSS Example usage isoverse/isoreader#145
