### 2020-12-14 80f0440
correct vignette and documentation links to iso_turn_info_messages_on/off
several links incorrectly had singular "message" instead of "messages" in the link (fixes #145)
