### 2020-12-14 isoverse/isoreader#145 @sebkopf
Thank you for catching this. The function has been renamed to `iso_turn_info_messages_off()` (plural "messages" instead of singular) but clearly not all links were updated. Should be fixed with the above commit.
