### 2020-12-14 c44dd96
add GPL2 license file
closes #141
