### 2020-12-14 9017ae7
update package authors
addresses #142
