+++ DESCRIPTION
@@ -2,7 +2,23 @@ Package: isoreader
 Title: Read Stable Isotope Data Files
 Description: Interface to the raw data file formats commonly encountered in scientific disciplines that make use of stable isotopes.
 Version: 1.2.7
-Authors@R: person("Sebastian", "Kopf", email = "sebastian.kopf@colorado.edu", role = c("aut", "cre"), comment = c(ORCID = "0000-0002-2044-0201"))
+Authors@R: 
+    c(person(
+        given = "Sebastian", family = "Kopf", 
+        email = "sebastian.kopf@colorado.edu", 
+        role = c("aut", "cre"), 
+        comment = c(ORCID = "0000-0002-2044-0201")),
+    person(
+        given = "Brett", family = "Davidheiser-Kroll", 
+        email = "brett.davidheiserkroll@colorado.edu", 
+        role = c("aut"), 
+        comment = c(ORCID = "0000-0002-6153-7851")),
+    person(
+        given = "Ilja", family = "Kocken", 
+        email = "i.j.kocken@uu.nl", 
+        role = c("aut"), 
+        comment = c(ORCID = "0000-0003-2196-8718"))
+    )
 URL: https://github.com/isoverse/isoreader
 BugReports: https://github.com/isoverse/isoreader/issues
 Depends:
