### 2020-12-14 isoverse/isoreader#142 @sebkopf
thank you for catching this, description file updated with all co-authors
