### 2020-12-14 a0f1978
add package version constraints and renv
addresses #143
