+++ DESCRIPTION
@@ -30,22 +30,22 @@ Imports:
     magrittr,
     rlang (>= 0.4.5),
     tidyselect (>= 1.0.0),
-    vctrs (>= 0.2.4),
+    vctrs (>= 0.3.4),
     tibble (>= 3.0.0),
-    dplyr (>= 0.8.5),
+    dplyr (>= 1.0.0),
     tidyr (>= 1.0.0),
-    glue,
-    stringr,
-    purrr,
-    future,
-    lubridate,
-    readr,
-    progress,
+    glue (>= 1.4.0),
+    stringr (>= 1.4.0),
+    purrr (>= 0.3.4),
+    future (>= 1.18.0),
+    lubridate (>= 1.7.9),
+    readr (>= 1.4.0),
+    progress (>= 1.2.2),
     xml2 (>= 1.3.1),
-    rhdf5,
-    feather,
-    UNF,
-    openxlsx
+    rhdf5 (>= 2.34.0),
+    feather (>= 0.3.5),
+    UNF (>= 2.0.6),
+    openxlsx (>= 4.1.5)
 Suggests:
     testthat,
     readxl,
