### 2020-12-14 isoverse/isoreader#143 @sebkopf
Thank you @kdorheim , this is fantastic. we stopped using packrat a while back because it caused more trouble than it solved but `renv` looks like it does a great job with versioning --> added to the repo.
