### 2020-12-14 7f2004c
update R depenceny to at least R 3.6 which changed the behaviour of sample
#144
