### 2020-12-14 6142257
add devtools to suggestions so it's part of renv
