### 2020-12-14 f523f30
update R CMD check github action to use latest packages instead of renv
