### 2020-12-14 58401a1
update github actions
use shell instead of R script to simplify and reduce run time
