+++ .github/workflows/R-CMD-check.yaml
@@ -38,11 +38,8 @@ jobs:
 
       - uses: r-lib/actions/setup-pandoc@master
 
-      - name: Init
-        run: |
-          # don't load .Rprofile (i.e. don't use renv, use latest packages)
-          unlink(".Rprofile")
-        shell: Rscript {0}
+      - name: Remove .Rprofile
+        run: rm -f .Rprofile
 
       - name: Query dependencies
         run: |
