### 2020-12-14 isoverse/isoreader#144 @sebkopf
Some of the warnings may be version related, especially with the [change in `sample` from R 3.6 onward](https://statisticaloddsandends.wordpress.com/2019/06/19/rs-sample-function-works-differently-from-r-3-6-0-onwards/). With `renv` set up now, would you be willing to run the following again to test?

1. `git clone --single-branch --branch joss https://github.com/isoverse/isoreader.git`
2. `renv::restore()`
3. `devtools::test()`

With this, the only warning should be the following if running interactively in RStudio:

**test-continuous-flow.R:117: warning: test that dxf files can be read
[ONE-TIME WARNING] Forked processing ('multicore') is disabled in future (>= 1.13.0) when running R from RStudio, because it is considered unstable. Because of this, plan("multicore") will fall back to plan("sequential"), and plan("multiprocess") will fall back to plan("multisession") - not plan("multicore") as in the past. For more details, how to control forked processing or not, and how to silence this warning in future R sessions, see ?future::supportsMulticore**
