### 2020-12-14 e58edb5
fix github actions issue for windows
windows rm command does not have the -f attribute
