### 2020-12-14 #2878 @sebkopf
thank you @kdorheim! pull request for the changes to address the raised issues is here: https://github.com/isoverse/isoreader/pull/148
