### 2020-12-15 isoverse/isoreader#144 @kdorheim
It looks like might be some information about the `Bioconductor` package may be missing

```
renv::restore()
Do you want to proceed? [y/N]: y
Retrieving 'https://mran.microsoft.com/snapshot/2020-04-24/bin/macosx/el-capitan/contrib/3.6/codetools_0.2-16.tgz' ...
	OK [file is up to date]
Error: failed to determine Bioconductor repositories
```
