### 2020-12-16 isoverse/isoreader#144 @sebkopf
Thank you for testing this. I'm not quite sure how to troubleshoot this error as the only attribute of the project's renv configuration that plays a role is the Bioconductor version specified in the [`renv.lock` lockfile (3.12)](https://github.com/isoverse/isoreader/blob/joss/renv.lock#L12), which is the correct version for the lock file's R version (4.0.2). The error is thrown by [the function that tries to find the BioConductor repository URLs](https://github.com/rstudio/renv/blob/master/R/bioconductor.R#L48) for this Bioconductor version which are normally just provided by the `BiocManager` package. Could you try `BiocManager::repositories()` and `BiocManager::repositories(version = "3.12")` and see if it provides repo URLs or a more informative error? I suspect this might be an R - BioConductor version conflict which would be unfortunate because it would require making the minimal R version for this package R 4.0 instead of 3.6
