### 2020-12-16 isoverse/isoreader#144 @kdorheim
Yes updating the R version 4 did the trick, which as mentioned is unfortunate.

devtools::test() passes with 71 warnings

```
OK:       1241
Failed:   0
Warnings: 71
Skipped:  0
```
all of the warnings look like they are related to

```
test-scan.R:53: warning: test that scn files can be read
Problem with `mutate()` input `result`.
ℹ Unrecognized time zone 'America/New_York'
ℹ Input `result` is `purrr::map2(...)`.
```
sorry didn't mean to make this more complicated than it already was 😬
