### 2020-12-17 73f2d50
requiring R version 4.0 for bioconductor compatibility
addressing an issue rasied in #144
