### 2020-12-17 isoverse/isoreader#144 @sebkopf
okay, R 4.0 it is, probably better for bioconductor support in the long run, several of the rhdf related packages are not backwards compatible.

As for the time zone warnings, seems like a problem with some of the lubridate calls but I could not reproduce it setting my system time zone to `America/New_York`. Perusing stack exchange there is some suggestion this could be a problem in lubridate that occurs in some operating system / R version combinations. Would you mind sharing your `sessionInfo()`? I can try to reproduce the warnings to resolve or at least catch them in a more informative way.
