### 2020-12-22 isoverse/isoreader#144 @kdorheim
Hi @sebkopf here is the session info for you

```
R version 4.0.3 (2020-10-10)
Platform: x86_64-apple-darwin17.0 (64-bit)
Running under: macOS Mojave 10.14.6

Matrix products: default
BLAS:   /System/Library/Frameworks/Accelerate.framework/Versions/A/Frameworks/vecLib.framework/Versions/A/libBLAS.dylib
LAPACK: /Library/Frameworks/R.framework/Versions/4.0/Resources/lib/libRlapack.dylib

locale:
[1] en_US.UTF-8/en_US.UTF-8/en_US.UTF-8/C/en_US.UTF-8/en_US.UTF-8

attached base packages:
[1] stats     graphics  grDevices datasets  utils     methods   base

other attached packages:
[1] feather_0.3.5   readxl_1.3.1    isoreader_1.2.7 testthat_2.3.2

 [1] Rcpp_1.0.5          lubridate_1.7.9     tidyr_1.1.1         listenv_0.8.0       prettyunits_1.1.1
 [6] ps_1.3.4            utf8_1.1.4          assertthat_0.2.1    rprojroot_1.3-2     digest_0.6.25
[11] R6_2.4.1            cellranger_1.1.0    backports_1.1.9     pillar_1.4.6        rlang_0.4.7
[16] progress_1.2.2      rstudioapi_0.11     callr_3.5.1         R.utils_2.9.2       R.oo_1.23.0
[21] desc_1.2.0          devtools_2.3.2      readr_1.4.0         stringr_1.4.0       compiler_4.0.3
[26] UNF_2.0.6           pkgconfig_2.0.3     base64enc_0.1-3     pkgbuild_1.1.0      globals_0.12.5
[31] tidyselect_1.1.0    tibble_3.0.3        codetools_0.2-16    fansi_0.4.1         future_1.18.0
[36] crayon_1.3.4        dplyr_1.0.2         withr_2.3.0         rhdf5filters_1.2.0  R.methodsS3_1.8.0
[41] lifecycle_0.2.0     magrittr_1.5        zip_2.1.0           cli_2.0.2           stringi_1.4.6
[46] renv_0.12.3         fs_1.4.2            remotes_2.2.0       xml2_1.3.2          ellipsis_0.3.1
[51] generics_0.0.2      vctrs_0.3.4         openxlsx_4.1.5      Rhdf5lib_1.12.0     tools_4.0.3
[56] glue_1.4.1          purrr_0.3.4         hms_0.5.3           processx_3.4.4      pkgload_1.1.0
[61] parallel_4.0.3      yaml_2.2.1          rhdf5_2.34.0        BiocManager_1.30.10 sessioninfo_1.1.1
[66] memoise_1.1.0       usethis_2.0.0
```

Output from sessionInfo and devtools::test()
[sessionInfo.txt](https://github.com/isoverse/isoreader/files/5731762/sessionInfo.txt)
[devtools_test.txt](https://github.com/isoverse/isoreader/files/5731767/devtools_test.txt)
