### 2020-12-31 isoverse/isoreader#144 @sebkopf
thanks @kdorheim, sorry this took me a while to figure out. I installed Mojave and R 4.0.3 but could not reproduce the error, turns out it was something introduced in R 4.0.2 that crated issues for some MacOS systems with the timezone database R uses (apparently dependent on how R was installed). I could finally reproduce it by deliberately setting `Sys.setenv(TZDIR = "internal")`. There's a detailed and quite fascinating discussion in the lubridate issue tracker on this issue: https://github.com/tidyverse/lubridate/issues/928

Anyways, long story short, it was fixed in the latest CRAN release of lubridate (1.7.9.2) so I updated the renv. Could you run the following one more time to test?

- `git clone --single-branch --branch joss https://github.com/isoverse/isoreader.git` or `git pull` if you still have the previous checkout
- `renv::restore()` (should updated `packageVersion("lubridate")` to 1.7.9.2)
- `devtools::test()`

Thank you!
