### 2021-01-04 af4bad2
update lubridate dependency to avoid timezone errors
