+++ DESCRIPTION
@@ -38,7 +38,7 @@ Imports:
     stringr (>= 1.4.0),
     purrr (>= 0.3.4),
     future (>= 1.18.0),
-    lubridate (>= 1.7.9),
+    lubridate (>= 1.7.9.2),
     readr (>= 1.4.0),
     progress (>= 1.2.2),
     xml2 (>= 1.3.1),
