### 2021-01-19 isoverse/isoreader#144 @kdorheim
@sebkopf sorry about the delay my availability has really become limited these days

no more time zone warning messages!!! This issue can be closed.
