### 2021-01-22 #2878 @r-barnes
For whatever reason, I'm unable to see the `LICENSE` file. However, CRAN can be a bear to work with and I see the license info in the `DESCRIPTION` file that CRAN uses, so I feel myself satisfied.
