### 2021-01-22 isoverse/isoreader#149 @r-barnes
Alphabetize list of supported types
Please alphabetize the list of supported types in the README to make it easier to determine what is supported (unless there are good reasons for the current organization).

openjournals/joss-reviews/issues/2878
