### 2021-01-22 isoverse/isoreader#150 @r-barnes
Add additional tests for supported file types
The following file types claim to be supported, but I don't see tests of this functionality:
```
.txt
.di.rda
.di.rds
.iarc
.cf.rda
.cf.rds
.scn
```
openjournals/joss-reviews/issues/2878
