### 2021-01-22 isoverse/isoreader#151 @r-barnes
Trouble installing
Trying `install.packages('isoreader')` ultimately gives:
```
ERROR: dependencies ‘rhdf5’, ‘feather’ are not available for package ‘isoreader’
* removing ‘/home/rick/R/x86_64-pc-linux-gnu-library/4.0/isoreader’
```
I haven't found `rhdf5` in CRAN. Are there other steps I'm supposed to take for installation?

openjournals/joss-reviews/issues/2878
