### 2021-01-22 isoverse/isoreader#152 @r-barnes
Paper missing references
The paper at https://github.com/isoverse/isoreader/blob/master/paper.md is missing references

openjournals/joss-reviews/issues/2878
