### 2021-01-22 #2878 @r-barnes
Sorry I'm so behind on this!

I've just finished looking through the repo. In general, things look pretty good. I've opened a few issues, as above. I reiterate the main points below:

1. Installation did not complete for me. It looks like some dependencies are not on CRAN and I didn't immediately see instructions for overcoming this.
2. This isn't my field, so I don't have appropriate input data available, but it looks highly plausible that the software does what is claimed. In particular, the impressively thorough [documentation](https://isoreader.isoverse.org/reference/) and the presence of unit tests are good evidence. However, the authors do claim the software will read a number of file types, but only have tests for a subset of these. From a testing perspective alone, it would be good to fill this gap. In this case, it's also important for evaluating functionality (at least for me).
3. Example usage. If I click around I can find a lot of complex examples. I think the README itself would be facilitated by having an extremely simple example to get the reader started (similar to what I do [here](https://github.com/r-barnes/dggridR), perhaps). I feel like a simple example helps sell viewers on the idea of looking in depth.

I think once these are taken care of, this should be good to go.
