### 2021-01-26 #2878 @sebkopf
Thank you @r-barnes for going through this. A few quick questions for clarification

> For whatever reason, I'm unable to see the LICENSE file. However, CRAN can be a bear to work with and I see the license info in the DESCRIPTION file that CRAN uses, so I feel myself satisfied.

Thank you for checking. The `LICENSE` is in the Joss branch at https://github.com/isoverse/isoreader/tree/joss, waiting for all edits from the review process to merge.

> 1. Installation did not complete for me. It looks like some dependencies are not on CRAN and I didn't immediately see instructions for overcoming this.

Could you copy which dependencies it is having trouble with? I'm testing on Windows, Linux and Mac OS with GitHub actions but could not find issues with dependencies except for the unreleased R development version (4.1) where `devtools` is not yet available. I suspect it might be `digest` which has in the past sometimes been a problem on windows (instructions for reinstall in the readme) or the bioconductor dependency `rhdf5` if `BiocManager` is not pulled correctly from CRAN. We can definitely add some troubleshooting guidelines for these if they are what is causing the problem.

> 2. This isn't my field, so I don't have appropriate input data available, but it looks highly plausible that the software does what is claimed. In particular, the impressively thorough documentation and the presence of unit tests are good evidence. However, the authors do claim the software will read a number of file types, but only have tests for a subset of these. From a testing perspective alone, it would be good to fill this gap. In this case, it's also important for evaluating functionality (at least for me).

The unit tests for the different file types (continuous flow, dual inlet and scan data) are in different test files to keep them easier to maintain with several example files included for all supported data types (for exactly the reason you mentioned, important to ensure functionality across the board):
 - https://github.com/isoverse/isoreader/blob/master/tests/testthat/test-continuous-flow.R for testing .iarc, .dxf, .cf files
 - https://github.com/isoverse/isoreader/blob/master/tests/testthat/test-dual-inlet.R for testing .did, .caf, .txt files
 - https://github.com/isoverse/isoreader/blob/master/tests/testthat/test-scan.R for testing .scn files
 - https://github.com/isoverse/isoreader/blob/master/tests/testthat/test-export.R for testing export and re-import of .cf.rds, .did.rds and .scn.rds

Note that not all of the example files are included on CRAN because of the limits on package file size but are always tested as part of the GitHub continuous integration. Also, some file extensions have more test files than others because we found more variability in the format (despite identical extension). Would you recommend additional tests or are these sufficient?

> 3. Example usage. If I click around I can find a lot of complex examples. I think the README itself would be facilitated by having an extremely simple example to get the reader started (similar to what I do here, perhaps). I feel like a simple example helps sell viewers on the idea of looking in depth.

We originally had this in the README but it got too cluttered so decided to put it into a `Quick Start` vignette, which is at https://isoreader.isoverse.org/articles/quick_start.html and covers some very basic code usage for the package. Would you suggest including this in the README as well or would it suffice to make a clear `How to get started` section in the readme that links to this vignette? (I realize we completely forgot to link this in the readme!)

Thank you very much for your time!
