### 2021-02-03 6e66b33
split release and devel R-CMD check actions
