+++ .github/workflows/R-CMD-check-devel.yaml
@@ -8,10 +8,10 @@ on:
   schedule:
     - cron: "0 0 * * FRI"
 
-name: R-CMD-check
+name: devel
 
 jobs:
-  R-CMD-check:
+  devel:
     runs-on: ${{ matrix.config.os }}
 
     name: ${{ matrix.config.os }} (${{ matrix.config.r }})
