+++ pkgdown.yml
@@ -1,6 +1,6 @@
 pandoc: 2.7.3
 pkgdown: 1.6.1.9000
-pkgdown_sha: 6094ac3dd80a1f3f1d4e2f47d0ed4716d51362b0
+pkgdown_sha: 84f234cf19153ce99bf1bcf875dc56248ddf0cbe
 articles:
   continuous_flow: continuous_flow.html
   development: development.html
@@ -8,7 +8,7 @@ articles:
   operations: operations.html
   quick_start: quick_start.html
   scan: scan.html
-last_built: 2020-11-24T21:40Z
+last_built: 2021-02-03T03:13Z
 urls:
   reference: https://isoreader.isoverse.org//reference
   article: https://isoreader.isoverse.org//articles
