### 2021-02-03 #2878 @sebkopf
Hi @r-barnes, just a quick note that the README updates (and all other JOSS review related updates) are now merged into master and reflected at https://isoreader.isoverse.org/ (including clear links to the vignettes and a snippet of example code). Let me know if you have any other questions regarding the testing details mentioned above or if this all sounds good for moving towards publication.

Thank you very much.
