### 2021-02-13 f48d17b
check vignette compatibility without pandoc
