+++ .github/workflows/R-CMD-check.yaml
@@ -35,8 +35,6 @@ jobs:
         with:
           r-version: ${{ matrix.config.r }}
 
-      - uses: r-lib/actions/setup-pandoc@master
-
       - name: Remove .Rprofile
         run: rm .Rprofile
 
