### 2021-02-13 0bc7d26
switch from paged_table to kable
