+++ vignettes/continuous_flow.Rmd
@@ -40,7 +40,7 @@ Reading continuous flow files is as simple as passing one or multiple file or fo
 
 ```{r, message = FALSE}
 # all available examples
-iso_get_reader_examples() %>% rmarkdown::paged_table()
+iso_get_reader_examples() %>% knitr::kable()
 ```
 
 ```{r}
@@ -58,7 +58,7 @@ cf_files <-
 The `cf_files` variable now contains a set of isoreader objects, one for each file. Take a look at what information was retrieved from the files using the `iso_get_data_summary()` function.
 
 ```{r}
-cf_files %>% iso_get_data_summary() %>% rmarkdown::paged_table()
+cf_files %>% iso_get_data_summary() %>% knitr::kable()
 ```
 
 ## Problems
@@ -66,8 +66,8 @@ cf_files %>% iso_get_data_summary() %>% rmarkdown::paged_table()
 In case there was any trouble with reading any of the files, the following functions provide an overview summary as well as details of all errors and warnings, respectively. The examples here contain no errors but if you run into any unexpected file read problems, please file a bug report in the [isoreader issue tracker](https://github.com/isoverse/isoreader/issues).
 
 ```{r}
-cf_files %>% iso_get_problems_summary() %>% rmarkdown::paged_table()
-cf_files %>% iso_get_problems() %>% rmarkdown::paged_table()
+cf_files %>% iso_get_problems_summary() %>% knitr::kable()
+cf_files %>% iso_get_problems() %>% knitr::kable()
 ```
 
 # File Information
@@ -76,7 +76,7 @@ Detailed file information can be aggregated for all isofiles using the `iso_get_
 
 ```{r}
 # all file information
-cf_files %>% iso_get_file_info(select = c(-file_root)) %>% rmarkdown::paged_table()
+cf_files %>% iso_get_file_info(select = c(-file_root)) %>% knitr::kable()
 # select file information
 cf_files %>%
   iso_get_file_info(
@@ -90,7 +90,7 @@ cf_files %>%
     ),
     # explicitly allow for file specific rename (for the new ID column)
     file_specific = TRUE
-  ) %>% rmarkdown::paged_table()
+  ) %>% knitr::kable()
 ```
 
 ## Select/Rename
@@ -109,7 +109,7 @@ cf_files2 <- cf_files %>%
   )
 
 # fetch all file info
-cf_files2 %>% iso_get_file_info() %>% rmarkdown::paged_table()
+cf_files2 %>% iso_get_file_info() %>% knitr::kable()
 ```
 
 ## Filter
@@ -120,13 +120,13 @@ Any collection of isofiles can also be filtered based on the available file info
 # find files that have 'acetanilide' in the new ID field
 cf_files2 %>% iso_filter_files(grepl("acetanilide", ID)) %>%
   iso_get_file_info() %>%
-  rmarkdown::paged_table()
+  knitr::kable()
 
 # find files that were run since 2015
 cf_files2 %>%
   iso_filter_files(`Date & Time` > "2015-01-01") %>%
   iso_get_file_info() %>%
-  rmarkdown::paged_table()
+  knitr::kable()
 ```
 
 ## Mutate
@@ -148,7 +148,7 @@ cf_files3 <-
 cf_files3 %>%
   iso_get_file_info() %>%
   iso_make_units_explicit() %>%
-  rmarkdown::paged_table()
+  knitr::kable()
 ```
 
 ## Add
@@ -171,13 +171,13 @@ new_info <-
        "6617_IAEA600",  "no",      "did not inject properly"
     )
   )
-new_info %>% rmarkdown::paged_table()
+new_info %>% knitr::kable()
 
[…]
