+++ vignettes/development.Rmd
@@ -41,7 +41,7 @@ new_reader <- function(ds, options = list()) {
 
 # register new reader
 readers <- iso_register_dual_inlet_file_reader(".new.did", "new_reader")
-rmarkdown::paged_table(readers)
+knitr::kable(readers)
 
 # copy an example file from the package with the new extension
 iso_get_reader_example("dual_inlet_example.did") %>% file.copy(to = "example.new.did")
