### 2021-02-13 5f0604b
fix issue in development vignette (kable does not support special characters)
