+++ vignettes/development.Rmd
@@ -103,15 +103,13 @@ This structure representation shows recognized control elements in `<...>` and d
 For an overview of all the control elements that are currently consider, use the internal `get_ctrl_blocks_config_df()` function.
 
 ```{r}
-isoreader:::get_ctrl_blocks_config_df() %>%
-  knitr::kable()
+isoreader:::get_ctrl_blocks_config_df()
 ```
 
 Additional information can be gleaned from the so-called control blocks, which are larger structural elements of Isodat binary files and are kept in a data frame within the binary object (again only available in debug mode).
 
 ```{r}
-bin$C_blocks %>%
-  knitr::kable()
+bin$C_blocks
 ```
 
 Same as for specific byte positions, one can use the control blocks to navigate the file and `map_binary_structure`.
