### 2021-02-15 444bc4b
remove .rda support
closes #153
