+++ tests/testthat/test-utils.R
@@ -288,7 +288,7 @@ test_that("test that get support file types are listed", {
   expect_true(is.data.frame(iso_get_supported_file_types()))
   expect_equal(
     iso_get_supported_file_types() %>% names(),
-    c("extension", "software", "description", "type", "call")
+    c("type", "extension", "software", "description", "call")
   )
 })
 
