+++ man/iso_export_to_excel.Rd
@@ -53,7 +53,6 @@ This function exports the passed in iso_files to Excel. The different kinds of d
 \seealso{
 Other export functions: 
 \code{\link{iso_export_to_feather}()},
-\code{\link{iso_export_to_rda}()},
 \code{\link{iso_save}()}
 }
 \concept{export functions}

[…]
