### 2021-02-15 c3afc76
update readme with missing .rda reader
