+++ README.md
@@ -4,7 +4,7 @@
 # isoreader <a href='https://isoreader.isoverse.org'><img src='man/figures/isoreader_logo_thumb.png' align="right" height="138.5"/></a>
 
 [![CRAN\_Status\_Badge](https://www.r-pkg.org/badges/version/isoreader)](https://cran.r-project.org/package=isoreader)
-[![Git\_Hub\_Version](https://img.shields.io/badge/GitHub-1.2.7-purple.svg?style=flat-square)](https://github.com/isoverse/isoreader/commits)
+[![Git\_Hub\_Version](https://img.shields.io/badge/GitHub-1.2.8-purple.svg?style=flat-square)](https://github.com/isoverse/isoreader/commits)
 [![Documentation](https://img.shields.io/badge/docs-online-green.svg)](https://isoreader.isoverse.org/)
 [![R build
 status](https://github.com/isoverse/isoreader/workflows/R-CMD-check/badge.svg)](https://github.com/isoverse/isoreader/actions?workflow=R-CMD-check)
@@ -89,7 +89,7 @@ iso_files <- iso_read_scan(data_folder)
 #> Info: reading file 'full_scan_example.scn' with '.scn' reader...
 #> Info: reading file 'peak_shape_scan_example.scn' with '.scn' reader...
 #> Info: reading file 'time_scan_example.scn' with '.scn' reader...
-#> Info: finished reading 4 files in 0.82 secs
+#> Info: finished reading 4 files in 1.21 secs
 
 iso_files
 #> Data from 4 scan iso files: 
@@ -108,18 +108,16 @@ Currently supported file types:
 
 | extension | software  | description                         | type            |
 | :-------- | :-------- | :---------------------------------- | :-------------- |
-| .did      | Isodat    | Dual Inlet file format (newer)      | dual inlet      |
-| .caf      | Isodat    | Dual Inlet file format (older)      | dual inlet      |
-| .txt      | Nu        | Dual Inlet file format              | dual inlet      |
-| .di.rda   | isoreader | R Data Archive (deprecated)         | dual inlet      |
-| .di.rds   | isoreader | R Data Storage                      | dual inlet      |
 | .cf       | Isodat    | Continuous Flow file format (older) | continuous flow |
+| .cf.rds   | isoreader | R Data Storage                      | continuous flow |
 | .dxf      | Isodat    | Continuous Flow file format (newer) | continuous flow |
 | .iarc     | ionOS     | Continuous Flow data archive        | continuous flow |
-| .cf.rda   | isoreader | R Data Archive (deprecated)         | continuous flow |
-| .cf.rds   | isoreader | R Data Storage                      | continuous flow |
-| .scn      | Isodat    | Scan file format                    | scan            |
+| .caf      | Isodat    | Dual Inlet file format (older)      | dual inlet      |
+| .di.rds   | isoreader | R Data Storage                      | dual inlet      |
+| .did      | Isodat    | Dual Inlet file format (newer)      | dual inlet      |
+| .txt      | Nu        | Dual Inlet file format              | dual inlet      |
 | .scan.rds | isoreader | R Data Storage                      | scan            |
+| .scn      | Isodat    | Scan file format                    | scan            |
 
 ## Documentation
 
