+++ .github/workflows/pkgdown.yaml
@@ -36,8 +36,7 @@ jobs:
         run: |
           install.packages("remotes")
           remotes::install_deps(dependencies = TRUE)
-          remotes::install_dev("pkgdown")
-          remotes::install_github("tidyverse/tidytemplate")
+          remotes::install_cran("pkgdown")
         shell: Rscript {0}
 
       - name: Install package
