+++ .github/workflows/pkgdown.yaml
@@ -49,7 +49,10 @@ jobs:
         shell: Rscript {0}
 
       - name: Install package
-        run: R CMD INSTALL .
+        run: |
+          remotes::install_cran("devtools")
+          devtools::install_github("isoverse/isoreader")
+        shell: Rscript {0}
 
       - name: Deploy package
         run: pkgdown::deploy_to_branch(new_process = FALSE)
