### 2021-02-15 isoverse/isoreader#150 @sebkopf
The unit tests for the different file types (continuous flow, dual inlet and scan data) are in different test files to keep them easier to maintain with several example files included for all supported data types (including the abovementioned .txt, .iarc, .scn and the formats generated by the package itself: .di.rds, .cf.rds, and .scan.rds). Please also note that the older R storage formats (.di.rda, .cf.rda) were used by earlier versions of this package for export but have been switched to .rds storage format. The support for .rda was deprecated before version 1.0 and has now been removed entirely (#153). Apologies if this has caused any confusion.

Test files for all supported file types:

https://github.com/isoverse/isoreader/blob/master/tests/testthat/test-continuous-flow.R for testing .iarc, .dxf, .cf files
https://github.com/isoverse/isoreader/blob/master/tests/testthat/test-dual-inlet.R for testing .did, .caf, .txt files
https://github.com/isoverse/isoreader/blob/master/tests/testthat/test-scan.R for testing .scn files
https://github.com/isoverse/isoreader/blob/master/tests/testthat/test-export.R for testing export and re-import of .cf.rds, .did.rds and .scn.rds

Note that not all of the example files are included on CRAN because of the limits on package file size but are always tested as part of the GitHub continuous integration. Also, some file extensions have more test files than others because we found more variability in the format (despite identical extension). This is especially true for the binary formats (.dxf, .cf, .caf, .did, .scn). The plain text and XML based formats (.txt and .iarc) have been much more consistent and does only included one representative example file each.
