+++ .github/workflows/pkgdown.yaml
@@ -6,7 +6,14 @@ name: pkgdown
 
 jobs:
   pkgdown:
-    runs-on: {os: ubuntu-16.04, r: 'release', rspm: "https://packagemanager.rstudio.com/cran/__linux__/xenial/latest"}
+    runs-on: ${{ matrix.config.os }}
+
+    strategy:
+      fail-fast: false
+      matrix:
+        config:
+          - {os: ubuntu-16.04, r: 'release', rspm: "https://packagemanager.rstudio.com/cran/__linux__/xenial/latest"}
+    
     steps:
       - uses: actions/checkout@v2
 
