+++ README.Rmd
@@ -67,8 +67,8 @@ Currently supported file types:
 
 ```{r, echo=FALSE, warning=FALSE, message=FALSE}
 library(isoreader)
-iso_get_supported_file_types() %>%
-  dplyr::select(extension, software, description, type) %>%
+iso_get_supported_file_types() %>% 
+  dplyr::select(-call) %>%
   knitr::kable()
 ```
 
