### 2021-02-16 isoverse/isoreader#149 @sebkopf
Good point, kept the grouping the same (file types `continuous flow`, `dual inlet` and `scan`) but put the groups in alphabetical order and the file extensions within each group also in alphabetical order. Moved the `type` column in the first position to clarify the grouping by file type. Update now live on the master branch `README` and at https://isoreader.isoverse.org/
