+++ R/export.R
@@ -2,7 +2,7 @@
 
 #' Export data to Excel
 #'
-#' This function exports the passed in iso_files to Excel. The different kinds of data (raw data, file info, methods info, etc.) are exported to separate tabs within the excel file. Use the various \code{include_...} parameters to specify what information to include. Note that in rare instances where vectorized data columns exist in the file information (e.g. measurement_info), they are concatenated with ', ' in the excel export.
+#' This function exports the passed in iso_files to Excel. The different kinds of data (raw data, file info, methods info, etc.) are exported to separate tabs within the excel file. Use the various \code{include_...} parameters to specify what information to include. Note that in rare instances where vectorized data columns exist in the file information (e.g. measurement_info), they are concatenated with ', ' in the excel export. Note that the openxlsx package required for this export is not installed automatically as part of isoreader. Please install it manually if missing using \code{install.packages("openxlsx")}.
 #'
 #' @inheritParams iso_save
 #' @inheritParams iso_get_all_data
@@ -20,6 +20,15 @@ iso_export_to_excel <- function(
   with_ratios = NULL,
   quiet = default(quiet)) {
 
+  # check for availability
+  if (!requireNamespace("openxlsx", quietly = TRUE)) {
+    warning(
+      "The 'openxlsx' package is required to export to this file format, please install it using the following command: install.packages(\"openxlsx\")",
+      immediate. = TRUE, call. = FALSE
+    )
+    return(invisible(iso_files))
+  }
+  
   # safety checks
   if(!iso_is_object(iso_files)) stop("can only export iso files or lists of iso files", call. = FALSE)
   export_iso_files <- iso_as_file_list(iso_files)
@@ -55,7 +64,7 @@ iso_export_to_excel <- function(
   )
 
   # make excel workbook
-  wb <- createWorkbook()
+  wb <- openxlsx::createWorkbook()
 
   # file info
   if ("file_info" %in% names(all_data)) {
@@ -97,7 +106,7 @@ iso_export_to_excel <- function(
     problems <- all_data %>% select(.data$file_id, .data$problems) %>% unnest(.data$problems)
     add_excel_sheet(wb, "problems", problems)
   }
-  saveWorkbook(wb, filepath, overwrite = TRUE)
+  openxlsx::saveWorkbook(wb, filepath, overwrite = TRUE)
 
   return(invisible(iso_files))
 }
@@ -109,15 +118,15 @@ iso_export_to_excel <- function(
 add_excel_sheet <- function(wb, sheet_name, ..., dbl_digits = 2, col_max_width = 75) {
 
   # sheet
-  addWorksheet(wb, sheet_name)
-  hs <- createStyle(textDecoration = "bold") # header style
+  openxlsx::addWorksheet(wb, sheet_name)
+  hs <- openxlsx::createStyle(textDecoration = "bold") # header style
 
   # data
   sheet_data_sets <- list(...)
   start_row <- 1L
   for (sheet_data in sheet_data_sets) {
     if (ncol(sheet_data) > 0) {
-      writeData(wb, sheet_name, sheet_data, startRow = start_row, headerStyle = hs)
+      openxlsx::writeData(wb, sheet_name, sheet_data, startRow = start_row, headerStyle = hs)
       int_cols <- which(purrr::map_lgl(sheet_data, is.integer))
       dbl_cols <- setdiff(which(purrr::map_lgl(sheet_data, is.numeric)), int_cols)
       if (dbl_digits < 1) {
@@ -127,7 +136,7 @@ add_excel_sheet <- function(wb, sheet_name, ..., dbl_digits = 2, col_max_width =
       # integer column formatting
       if (length(int_cols) > 0) {
         openxlsx::addStyle(
-          wb, sheet_name, style = createStyle(numFmt = "0"),
[…]
