+++ NAMESPACE
@@ -192,7 +192,6 @@ importFrom(dplyr,summarize)
 importFrom(dplyr,tally)
 importFrom(dplyr,ungroup)
 importFrom(dplyr,vars)
-importFrom(feather,write_feather)
 importFrom(future,availableCores)
 importFrom(future,future)
 importFrom(future,plan)
@@ -205,11 +204,6 @@ importFrom(lubridate,interval)
 importFrom(magrittr,"%>%")
 importFrom(methods,is)
 importFrom(methods,setOldClass)
-importFrom(openxlsx,addWorksheet)
-importFrom(openxlsx,createStyle)
-importFrom(openxlsx,createWorkbook)
-importFrom(openxlsx,saveWorkbook)
-importFrom(openxlsx,writeData)
 importFrom(purrr,is_empty)
 importFrom(purrr,map)
 importFrom(purrr,map2)
