+++ DESCRIPTION
@@ -1,7 +1,7 @@
 Package: isoreader
 Title: Read Stable Isotope Data Files
 Description: Interface to the raw data file formats commonly encountered in scientific disciplines that make use of stable isotopes.
-Version: 1.2.7
+Version: 1.2.8
 Authors@R: 
     c(person(
         given = "Sebastian", family = "Kopf", 
@@ -42,14 +42,14 @@ Imports:
     readr (>= 1.4.0),
     progress (>= 1.2.2),
     xml2 (>= 1.3.1),
-    rhdf5 (>= 2.34.0),
-    feather (>= 0.3.5),
-    UNF (>= 2.0.6),
-    openxlsx (>= 4.1.5)
+    rhdf5 (>= 2.0.0),
+    UNF (>= 2.0.6)
 Suggests:
     devtools,
     testthat,
-    readxl,
+    feather (>= 0.3.5),
+    readxl (>= 1.3.1),
+    openxlsx (>= 4.1.5),
     knitr,
     rmarkdown,
     covr
