### 2021-02-16 006e178
make xml2 and rhdf5 optional
closes #154
