+++ DESCRIPTION
@@ -41,8 +41,6 @@ Imports:
     lubridate (>= 1.7.9.2),
     readr (>= 1.4.0),
     progress (>= 1.2.2),
-    xml2 (>= 1.3.1),
-    rhdf5 (>= 2.0.0),
     UNF (>= 2.0.6)
 Suggests:
     devtools,
@@ -50,6 +48,8 @@ Suggests:
     feather (>= 0.3.5),
     readxl (>= 1.3.1),
     openxlsx (>= 4.1.5),
+    xml2 (>= 1.3.1),
+    rhdf5 (>= 2.0.0),
     knitr,
     rmarkdown,
     covr
