### 2021-02-16 isoverse/isoreader#151 @sebkopf
`rhdf5` is part of [bioconductor](https://bioconductor.org/packages/release/bioc/html/rhdf5.html) which is a separate package management system from CRAN that was originally designed for bioinformatics functionality but has become much larger since. As a bioconductor package, `rhdf5` can unfortunately not be installed directly from CRAN. The line `biocViews` in the `DESCRIPTION` file should make it possible for R to figure out that it needs to look at the bioconductor package list as well but it appears this does not always get resolved correctly on all operating systems.  The manual installation command for `rhdf5` is

```
if (!requireNamespace("BiocManager", quietly = TRUE))
    install.packages("BiocManager")
BiocManager::install("rhdf5")
```

However, your issue prompted us to simplify the dependencies and make rhdf5 and several other dependencies that are only used under specific circumstances not part of the dependency list but instead make them part of the suggested packages and check whether they are actually installed whenever needed as recommended by the CRAN guidelines (and if not yet installed, provide an informative message explaining how to install them).
