### 2021-02-16 isoverse/isoreader#152 @sebkopf
I'm really sorry @r-barnes, I could not figure out what references are missing. Did you mean that they are not included in the `paper.md` file? Is it okay to keep them in `paper.bib` at https://github.com/isoverse/isoreader/blob/master/paper.bib? The JOSS preview service incorporates the references from the bib file correctly, see https://whedon.theoj.org/preview?id=e12f09b279c4a7bd5d8a7f17 for the preview. Please kindly re-open if I missed what references you found missing.
