### 2021-02-16 ba613dd
simplify makefile
