+++ Makefile
@@ -7,13 +7,13 @@ all: docu check
 docu:
 	Rscript -e "devtools::document(roclets=c('rd', 'collate', 'namespace'))"
 
+# test package functionality without all example files (= as if on CRAN)
 check:
 	R -q -e "devtools::check(env_vars = c())"
 
-# test package functionality without all example files
 auto_test:
-	R -q -e "rm(list = ls()); options("isoreader.run_file_tests" = FALSE); testthat::auto_test_package()"
+	R -q -e "rm(list = ls()); Sys.setenv(NOT_CRAN = \"false\"); testthat::auto_test_package()"
 
-# test all example files
+# test with all example files (= as if not on CRAN)
 auto_test_all:
 	R -q -e "rm(list = ls()); testthat::auto_test_package()"
