### 2021-04-18 isoverse/isoreader#152 @r-barnes
I think there are human-readable ways of putting citations in markdown, see, for instance, https://rmarkdown.rstudio.com/authoring_bibliographies_and_citations.html. But let's forego that for this work.
