### 2021-04-27 #2878 @sebkopf
Hi @mikldk . Got zenodo finally figured out:

The release version is 1.3.0 and the archive DOI is 10.5281/zenodo.4724083:

[![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.4724083.svg)](https://doi.org/10.5281/zenodo.4724083)
