### 2021-04-27 #2878 @sebkopf
> 👉📄 [Download article proof](https://raw.githubusercontent.com/openjournals/joss-papers/joss.02878/joss.02878/10.21105.joss.02878.pdf) 📄 [View article proof on GitHub](https://github.com/openjournals/joss-papers/blob/joss.02878/joss.02878/10.21105.joss.02878.pdf) 📄 👈

@mikldk This all looks great, happy to approve this proof. Is there anything else you need me to do? As mentioned above, zenodo DOI is [10.5281/zenodo.4724083](https://zenodo.org/badge/latestdoi/90583005)
