### 2021-04-27 41aa91b
add markdown in suggests
to address https://github.com/yihui/knitr/issues/1864
