+++ DESCRIPTION
@@ -51,6 +51,7 @@ Suggests:
     xml2 (>= 1.3.1),
     rhdf5 (>= 2.0.0),
     knitr,
+    markdown,
     rmarkdown,
     covr
 biocViews: 
