### 2021-05-01 #2878 @Kevin-Mattheus-Moerman
@sebkopf I've also proofread the paper and I think all looks good. I will now proceed to accept this submission in JOSS. Congratulations!
